+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 2.1.0
 
+- Added opcode 0DD5 (get_platform).
 - Partially fixed 0ADD (spawn car like a cheat) for GTA3. Now opcode can spawn not only the tank. Still no support for models with id greater than 127.
 - Fixed 0ADB to return GXT label instead of car name (same as in Cleo SA).
 - Crash fix in 0AF4 read_string_from_ini_file. Fixed some internal data-type sizes.